add some typing
