+++ .gitignore
@@ -30,12 +30,15 @@
 !/viseron/*
 !/viseron/config
 !/viseron/config/*
-!/viseron/detectors
-!/viseron/detectors/*
-!/viseron/detectors/darknet
-!/viseron/detectors/darknet/*
-!/viseron/detectors/edgetpu
-!/viseron/detectors/edgetpu/*
+!/viseron/detector
+!/viseron/detector/*
+!/viseron/detector/darknet
+!/viseron/detector/darknet/*
+!/viseron/detector/edgetpu
+!/viseron/detector/edgetpu/*
+!/viseron/helpers
+!/viseron/helpers/*
+!/viseron/post_processors
 !/viseron/mqtt
 !/viseron/mqtt/*
 !/viseron/post_processors
