+++ viseron/data_stream.py
@@ -4,8 +4,9 @@ import logging
 import uuid
 from queue import Queue
 from threading import Thread
-from typing import Any, Callable, Dict, List, Union
+from typing import Any, Callable, Dict, Union
 
+from tornado.ioloop import IOLoop
 from tornado.queues import Queue as tornado_queue
 
 from viseron.helpers import pop_if_full
@@ -26,20 +27,20 @@ class DataStream:
     _wildcard_subscribers: Dict[str, Any] = {}
     _data_queue: Queue = Queue(maxsize=100)
 
-    def __init__(self, ioloop):
+    def __init__(self, ioloop: IOLoop) -> None:
         data_consumer = Thread(target=self.consume_data)
         data_consumer.start()
         self.ioloop = ioloop
 
     @staticmethod
-    def publish_data(data_topic, data):
+    def publish_data(data_topic: str, data: Any) -> None:
         """Publish data to topic."""
         # LOGGER.debug(f"Publishing to data topic {data_topic}, {data}")
         DataStream._data_queue.put({"data_topic": data_topic, "data": data})
 
     @staticmethod
     def subscribe_data(
-        data_topic, callback: Union[Callable, Queue, tornado_queue]
+        data_topic: str, callback: Union[Callable, Queue, tornado_queue]
     ) -> uuid.UUID:
         """Subscribe to data on a topic.
         Returns a Unique ID which can be used to unsubscribe later."""
@@ -56,7 +57,7 @@ class DataStream:
         return unique_id
 
     @staticmethod
-    def unsubscribe_data(data_topic, unique_id: uuid.UUID):
+    def unsubscribe_data(data_topic: str, unique_id: uuid.UUID) -> None:
         """Unsubscribe from a topic using the Unique ID returned from subscribe_data."""
         LOGGER.debug(f"Unsubscribing from data topic {data_topic}, {unique_id}")
         if "*" in data_topic:
@@ -66,8 +67,10 @@ class DataStream:
         DataStream._subscribers[data_topic].pop(unique_id)
 
     def run_callbacks(
-        self, callbacks: List[Union[Callable, Queue, tornado_queue]], data
-    ):
+        self,
+        callbacks: Dict[uuid.UUID, Union[Callable, Queue, tornado_queue]],
+        data: Any,
+    ) -> None:
         """Run callbacks or put to queues."""
         for callback in callbacks.values():
             if callable(callback):
@@ -89,13 +92,13 @@ class DataStream:
                 f"Needs to be of type Callable or Queue, got {type(callback)}"
             )
 
-    def static_subscriptions(self, data_item):
+    def static_subscriptions(self, data_item: Dict[str, Any]) -> None:
         """Run callbacks for static subscriptions."""
         self.run_callbacks(
             DataStream._subscribers.get(data_item["data_topic"], {}), data_item["data"]
         )
 
-    def wildcard_subscriptions(self, data_item):
+    def wildcard_subscriptions(self, data_item: Dict[str, Any]) -> None:
         """Run callbacks for wildcard subscriptions."""
         for data_topic, callbacks in DataStream._wildcard_subscribers.items():
             if fnmatch.fnmatch(data_item["data_topic"], data_topic):
@@ -106,7 +109,7 @@ class DataStream:
 
                 self.run_callbacks(callbacks, data_item["data"])
 
-    def consume_data(self):
+    def consume_data(self) -> None:
         """Publish data to topics."""
         while True:
             data_item = self._data_queue.get()
