+++ viseron/detector/darknet/defaults.py
