+++ viseron/detector/edgetpu/defaults.py
